+++ compiler/tests-gen/org/jetbrains/kotlin/codegen/ir/ComposeLikeIrBytecodeTextTestGenerated.java
@@ -17,7 +17,7 @@ import java.util.regex.Pattern;
 
 /** This class is generated by {@link org.jetbrains.kotlin.generators.tests.TestsPackage}. DO NOT MODIFY MANUALLY */
 @SuppressWarnings("all")
-@TestMetadata("compiler/testData/codegen/composeLike")
+@TestMetadata("compiler/testData/codegen/composeLikeBytecodeText")
 @TestDataPath("$PROJECT_ROOT")
 @RunWith(JUnit3RunnerWithInners.class)
 public class ComposeLikeIrBytecodeTextTestGenerated extends AbstractComposeLikeIrBytecodeTextTest {
@@ -25,22 +25,22 @@ public class ComposeLikeIrBytecodeTextTestGenerated extends AbstractComposeLikeI
         KotlinTestUtils.runTest(this::doTest, TargetBackend.JVM_IR, testDataFilePath);
     }
 
-    public void testAllFilesPresentInComposeLike() throws Exception {
-        KotlinTestUtils.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler/testData/codegen/composeLike"), Pattern.compile("^(.+)\\.kt$"), null, TargetBackend.JVM_IR, true);
+    public void testAllFilesPresentInComposeLikeBytecodeText() throws Exception {
+        KotlinTestUtils.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler/testData/codegen/composeLikeBytecodeText"), Pattern.compile("^(.+)\\.kt$"), null, TargetBackend.JVM_IR, true);
     }
 
     @TestMetadata("default.kt")
     public void testDefault() throws Exception {
-        runTest("compiler/testData/codegen/composeLike/default.kt");
+        runTest("compiler/testData/codegen/composeLikeBytecodeText/default.kt");
     }
 
     @TestMetadata("defaultInline.kt")
     public void testDefaultInline() throws Exception {
-        runTest("compiler/testData/codegen/composeLike/defaultInline.kt");
+        runTest("compiler/testData/codegen/composeLikeBytecodeText/defaultInline.kt");
     }
 
     @TestMetadata("defaultLocal.kt")
     public void testDefaultLocal() throws Exception {
-        runTest("compiler/testData/codegen/composeLike/defaultLocal.kt");
+        runTest("compiler/testData/codegen/composeLikeBytecodeText/defaultLocal.kt");
     }
 }
